requirements.txt
Django==5.2.9
requests
dj-database-url==2.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
sentry-sdk[django]
python-dateutil
poweradmin @ git+https://bitbucket.org/josir/poweradmin.git@django52
